+++ ui/components/findings/table/column-findings.tsx
@@ -71,7 +71,9 @@ const FindingDetailsCell = ({ row }: { row: any }) => {
   return (
     <div className="flex max-w-10 justify-center">
       <TriggerSheet
-        triggerComponent={<InfoIcon className="text-primary" size={16} />}
+        triggerComponent={
+          <InfoIcon className="text-button-primary" size={16} />
+        }
         title="Finding Details"
         description="View the finding details"
         defaultOpen={isOpen}